refactor: fixed tests
