+++ src/__test__/Header.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Header } from "../Header";
+import { Header } from "../components/Common/Header/Header";
 import { render } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
