+++ src/__test__/Map.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Provider } from "react-redux";
-import { Map } from "../Map";
+import { Map } from "../pages/Map/Map";
 import { render } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
 import store from "../redux/store";
